# docs/swagger.yaml
basePath: /v1
definitions:
...
  api.CreateJobRequest:
    properties:
      input:
        additionalProperties: true
        type: object
      meta:
        additionalProperties: true
        type: object
      workflow:
        type: string
    type: object
  api.CreateJobResponse:
    properties:
      id:
        type: string
    type: object
...
  api.Job:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      id:
        type: string
      input:
        additionalProperties: true
        type: object
      meta:
        additionalProperties: true
        type: object
      startedAt:
        type: string
      status:
        $ref: '#/definitions/api.JobStatus'
      updatedAt:
        type: string
      workflow:
        type: string
    type: object
  api.JobListResponse:
    properties:
      cursor:
        type: string
      hasMore:
        type: boolean
      jobs:
        items:
          $ref: '#/definitions/api.Job'
        type: array
    type: object
  api.JobLogsResponse:
    properties:
      jobId:
        type: string
      logs:
        items:
          $ref: '#/definitions/api.LogEntry'
        type: array
    type: object
  api.JobResult:
    properties:
      artifacts:
        description: artifact IDs
        items:
          type: string
        type: array
      completedAt:
        type: string
      jobId:
        type: string
      result:
        additionalProperties: true
        type: object
      status:
        $ref: '#/definitions/api.JobStatus'
    type: object
  api.JobStatus:
    enum:
    - queued
    - running
    - succeeded
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - JobStatusQueued
    - JobStatusRunning
    - JobStatusSucceeded
    - JobStatusFailed
    - JobStatusCancelled
  api.JobStep:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      id:
        type: string
      input:
        additionalProperties: true
        type: object
      jobId:
        type: string
      name:
        type: string
      output:
        additionalProperties: true
        type: object
      startedAt:
        type: string
      status:
        $ref: '#/definitions/api.StepStatus'
      updatedAt:
        type: string
    type: object
  api.LogEntry:
    properties:
      level:
        type: string
      message:
        type: string
      stepId:
        type: string
      timestamp:
        type: string
    type: object
...
  api.QueueItem:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      data:
        additionalProperties: true
        type: object
      id:
        type: string
      jobId:
        type: string
      leasedAt:
        type: string
      state:
        $ref: '#/definitions/api.QueueState'
      updatedAt:
        type: string
    type: object
  api.QueueItemListResponse:
    properties:
      cursor:
        type: string
      hasMore:
        type: boolean
      items:
        items:
          $ref: '#/definitions/api.QueueItem'
        type: array
    type: object
  api.QueueState:
    enum:
    - pending
    - leased
    - done
    - dead
    type: string
    x-enum-varnames:
    - QueueStatePending
    - QueueStateLeased
    - QueueStateDone
    - QueueStateDead
...
  api.RequeueRequest:
    properties:
      jobId:
        type: string
      reason:
        type: string
    type: object
...
  api.StepStatus:
    enum:
    - pending
    - running
    - succeeded
    - failed
    - skipped
    type: string
    x-enum-varnames:
    - StepStatusPending
    - StepStatusRunning
    - StepStatusSucceeded
    - StepStatusFailed
    - StepStatusSkipped
...
host: localhost:3333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for the Agent Project Manager daemon. Orchestrates agent-based
    LLM workflows.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Agent Project Manager API
  version: "1.0"
paths:
...
  /jobs:
    get:
      consumes:
      - application/json
      description: Get a paginated list of jobs with optional filtering
      parameters:
      - description: Maximum number of jobs to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      - description: Filter by status (queued|running|succeeded|failed)
        in: query
        name: status
        type: string
      - description: Filter by workflow name
        in: query
        name: workflow
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JobListResponse'
      summary: List jobs
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Submit a new job to be processed
      parameters:
      - description: Job creation request
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/api.CreateJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.CreateJobResponse'
        "400":
          description: Invalid request body
          schema:
            type: string
      summary: Create a new job
      tags:
      - jobs
  /jobs/{jobId}:
    delete:
      consumes:
      - application/json
      description: Cancel a running or queued job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "404":
          description: Job not found
          schema:
            type: string
      summary: Cancel a job
      tags:
      - jobs
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Job'
        "404":
          description: Job not found
          schema:
            type: string
      summary: Get job details
      tags:
      - jobs
  /jobs/{jobId}/events:
    get:
      consumes:
      - application/json
      description: Stream job status updates via Server-Sent Events (SSE)
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Event stream
          schema:
            type: string
      summary: Stream job events
      tags:
      - jobs
  /jobs/{jobId}/logs:
    get:
      consumes:
      - application/json
      description: Get aggregated logs for a job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Number of log lines to return
        in: query
        name: tail
        type: integer
      - description: RFC3339 timestamp to filter logs from
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JobLogsResponse'
        "404":
          description: Job not found
          schema:
            type: string
      summary: Get job logs
      tags:
      - jobs
  /jobs/{jobId}/result:
    get:
      consumes:
      - application/json
      description: Get the latest result summary for a job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JobResult'
        "404":
          description: Job not found
          schema:
            type: string
      summary: Get job result
      tags:
      - jobs
  /jobs/{jobId}/retry:
    post:
      consumes:
      - application/json
      description: Retry a failed or cancelled job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.CreateJobResponse'
        "404":
          description: Job not found
          schema:
            type: string
      summary: Retry a job
      tags:
      - jobs
  /jobs/{jobId}/steps:
    get:
      consumes:
      - application/json
      description: Get a list of workflow steps for a job
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.JobStep'
            type: array
        "404":
          description: Job not found
          schema:
            type: string
      summary: List job steps
      tags:
      - jobs
  /jobs/{jobId}/steps/{stepId}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a workflow step
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Step ID
        in: path
        name: stepId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JobStep'
        "404":
          description: Step not found
          schema:
            type: string
      summary: Get step details
      tags:
      - jobs
  /jobs/{jobId}/steps/{stepId}/logs:
    get:
      consumes:
      - application/json
      description: Get logs for a specific workflow step
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Step ID
        in: path
        name: stepId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.JobLogsResponse'
        "404":
          description: Step not found
          schema:
            type: string
      summary: Get step logs
      tags:
      - jobs
...
  /queue/items:
    get:
      consumes:
      - application/json
      description: Get a paginated list of queue items with optional filtering
      parameters:
      - description: Maximum number of items to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      - description: Filter by state (pending|leased|done|dead)
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.QueueItemListResponse'
      summary: List queue items
      tags:
      - queue
  /queue/requeue:
    post:
      consumes:
      - application/json
      description: Requeue a job in the queue
      parameters:
      - description: Requeue request
        in: body
        name: requeue
        required: true
        schema:
          $ref: '#/definitions/api.RequeueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Requeue a job
      tags:
      - queue
...
schemes:
- http
- https
swagger: "2.0"