# internal/api/handlers_queue.go
package api

import (
	"encoding/json"
	"net/http"
)

// handleGetQueue handles GET /queue
// @Summary      Get queue statistics
// @Description  Get queue statistics and metrics
// @Tags         queue
// @Accept       json
// @Produce      json
// @Success      200  {object}  QueueStats
// @Router       /queue [get]
func handleGetQueue(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement queue stats retrieval
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(map[string]interface{}{})
}

// handleListQueueItems handles GET /queue/items
// @Summary      List queue items
// @Description  Get a paginated list of queue items with optional filtering
// @Tags         queue
// @Accept       json
// @Produce      json
// @Param        limit   query     int     false  "Maximum number of items to return"
// @Param        cursor  query     string  false  "Cursor for pagination"
// @Param        state   query     string  false  "Filter by state (pending|leased|done|dead)"
// @Success      200     {object}  QueueItemListResponse
// @Router       /queue/items [get]
func handleListQueueItems(w http.ResponseWriter, r *http.Request) {
	// TODO: Parse query parameters: limit, cursor, state
	// limit := r.URL.Query().Get("limit")
	// cursor := r.URL.Query().Get("cursor")
	// state := r.URL.Query().Get("state") // pending|leased|done|dead

	// TODO: Implement queue items listing logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode([]interface{}{})
}

// handleRequeue handles POST /queue/requeue
// @Summary      Requeue a job
// @Description  Requeue a job in the queue
// @Tags         queue
// @Accept       json
// @Produce      json
// @Param        requeue  body      RequeueRequest  true  "Requeue request"
// @Success      200      {string}  string  "OK"
// @Failure      400      {string}  string  "Invalid request"
// @Router       /queue/requeue [post]
func handleRequeue(w http.ResponseWriter, r *http.Request) {
	var req RequeueRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// TODO: Implement requeue logic
	_ = req

	w.WriteHeader(http.StatusOK)
}